+++ packages/tui/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Tool cards can now show passive model context as one sanitized, dim line without exposing it as a separate transcript message ([#13269](https://github.com/can1357/oh-my-pi/pull/13269) by [@andrebrait](https://github.com/andrebrait)).
+
 ## [18.3.1] - 2026-09-25
 
 ### Breaking Changes